Finished zoom and moving the field arround.
