+++ .vscode/settings.json
@@ -1,0 +1,9 @@
+{
+    "makefile.launchConfigurations": [
+        {
+            "cwd": "/home/paul/Projects/conway/build",
+            "binaryPath": "/home/paul/Projects/conway/build/conway",
+            "binaryArgs": []
+        }
+    ]
+}
